requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cyvcf2==0.7.0
dicttoxml==1.7.4
Django==1.10.5
django-audit-log==0.7.0
django-auditlog==0.4.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-discover-runner==1.0
django-filter==1.1.0
django-jsonfield==1.0.1
django-role-permissions==2.1.0
djangorestframework==3.6.4
enstaller==4.8.12
gunicorn==19.7.1
idna==2.6
interop==1.0.25
mysqlclient==1.3.12
myvariant==0.3.1
numpy==1.12.1
pysam==0.10.0
requests==2.18.4
scipy==1.0.0
six==1.11.0
sqlparse==0.2.3
urllib3==1.22
yolk==0.4.3
